trix-erp/requirements.txt
Flask==2.3.2
flask-login
Flask-Cors==4.0.0
SQLAlchemy
Flask-SQLAlchemy
flask-wtf
gunicorn==20.1.0
requests==2.31.0
Werkzeug==2.3.6
python-dotenv==1.0.0  # Opcional, si utilizas un archivo .env
python-docx==1.1.0
psycopg2-binary==2.9.7
PuLP
joblib==1.3.2
numpy==1.24.4
pandas
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.0
pytz==2023.3.post1
Pillow==10.1.0
plotly==5.18.0
haversine==2.5.1
folium==0.14.0
networkx==2.6.3
geopy==2.4.0
flask-compress




